systemd_%.bbappend: avoid overwriting systemd's default PACKAGECONFIG
Using += instead of _append accidentally disabled the default
PACKAGECONFIG, because += is evaluated first, and then the ??=
default is not used anymore.

One effect was that logging in triggered a "System is booting up. See
pam_nologin(8)" warning (for root) or even prevented access entirely
(non-root users).

While at it, also remove the redundant iptables dependency, which is
already covered by the systemd_228.bb PACKAGECONFIG[iptc].

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>