systemd_%.bbappend: try to make systemd-networkd socket-activated.
Try to make systemd-networkd socket-activated instead of getting
started unconditionally. This is the officially preferred way of
running it.

To get socket-activation turned on, we separate systemd-networkd
to a subpackage of its own, move the binary, the systemd service
and socket files to the new subpackage, advertise the subpackage
as one containing systemd-activated components, and override its
service file with the socket file.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>