connman: patch connman default blacklist and ip_forward toggling.
Added two small connman patches. One simply adds ve- and vb-,
the virtual ethernet and bridge interface prefixes, to the
default list of blacklisted interfaces.

The other saves the current global IP forwarding setting (from
/proc/sys/net/ipv4/ip_forward) when ConnMan is about to touch
it for the first time and restores it instead of resetting it
when ConnMan itself does not need IP forwarding any more. This
is far from perfect but not as broken as the original behavior
which simply assumed only ConnMan can need IP forwarding. Now
if we enable it before ConnMan is started, it will properly stay
on.

An alternative would be to patch the kernel procfs forward
control interface to support reference counting. +1 would increase
the reference and -1 would decrease it. 1 would be an alias to +1,
but 0 would still reset it to 0.

A third, and probably the best, alternative would be to check if we
could switch ConnMan's logic to control the per-interface forwarding
settings (/proc/sys/net/ipv4/conf/<ifname>/forwarding) instead of
controlling the setting globally.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>