ostro-app.bbclass: crude support for in-manifest path relocation.
Added crude support for in-manifest path relocation. If we find a
manifest.in for an application instead of a manifest, we process
it substituting @OSTRO_APP_ROOT@ with its value. This allows one
to write a working manifest(.in) for a containerless application
even if the command line to start the application needs to refer
to the relocated application path.

One way to improve this and make it much more flexible is to add
support for variable substitution for the loaded data (manifest)
before/during (service) template processing.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>