iot-app-fw.bb: install an applications.target for autostarting.
Install an applications.target which is used to collect all
applications that should be automatically started on boot.
The service generator produces an applications.target.wants
directory and populates it with the necessary symlinks.
Applications.target is always enabled, hence if autostarted
applications are installed, this should get then bootstrapped.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>