iot-app-fw_git.bb: update to latest version.
Updated to latest service generator. This adds support for
starting/running the service generator also inside a container
to prepare/start services (agentful mode).

Added/updated templates for both agentful and agentless mode
according to the new template format.

Added a manifest preprocessor to convert manifests from the
old to the new format on the fly.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>